requirements.txt
asgiref==3.7.2
crispy-bootstrap4==2024.1
Django==5.0.3
django-crispy-forms==2.1
django-grappelli==3.0.8
django-model-utils==4.4.0
django-notifications-hq==1.8.3
django-phonenumber-field==7.3.0
django-tables2==2.7.0
jsonfield==3.1.0
phonenumbers==8.13.32
pillow==10.2.0
pytz==2024.1
setuptools==69.2.0
sqlparse==0.4.4
swapper==1.3.0
tzdata==2024.1
